agents/summarizer_agent/test_agent.py
import unittest
from unittest.mock import Mock, patch
from agents.summarizer_agent.agent import TextSummarizerAgent, TextSummarizeInput, TextSummarizeOutput
from llm_provider.factory import LLMFactory
from common.message_bus import MessageBus
from logging import Logger

class TestTextSummarizerAgent(unittest.TestCase):
    def setUp(self):
        self.logger = Mock(Logger)
        self.llm_factory = Mock(LLMFactory)
        self.message_bus = Mock(MessageBus)
        self.agent_id = "text_summarizer_agent"
        self.config = {
            "chunk_size": 500,
            "max_summary_length": 100,
            "accuracy_threshold": 0.8,
            "completeness_threshold": 0.8,
            "relevance_threshold": 0.8
        }
        self.agent = TextSummarizerAgent(self.logger, self.llm_factory, self.message_bus, self.agent_id, self.config)
        self.agent.setup_graph() 

    def test_run(self):
        # Couldnt get this to work so just implemented a direct integ test use that instead
        pass
        
    def test_format_input(self):
        instruction = "Summarize this text"
        history = ["This is some maybe relevant history", "This is a sample text to summarize."]
        max_summary_length = 100
        expected_input = TextSummarizeInput(prompt=instruction, history=history, config={"max_summary_length": max_summary_length})

        input_data = self.agent._format_input({"prompt": instruction, "history": history, "agent_config": {"max_summary_length": max_summary_length}})

        self.assertEqual(input_data, expected_input)

    def test_calculate_accuracy_scores(self):
        summary = "Summary text"
        original_text = "This is a sample text to summarize."
        expected_accuracy_score = 0.9
        expected_completeness_score = 0.9

        accuracy_score, completeness_score = self.agent.calculate_accuracy_scores(summary, original_text)

        self.assertEqual(accuracy_score, expected_accuracy_score)
        self.assertEqual(completeness_score, expected_completeness_score)

    def test_calculate_relevance_score(self):
        summary = "Summary text"
        original_text = "This is a sample text to summarize."
        expected_relevance_score = 0.5  # Assuming a dummy value for the test

        with patch("agents.summarizer_agent.agent.cosine_similarity") as mock_cosine_similarity:
            mock_cosine_similarity.return_value = [[expected_relevance_score]]
            relevance_score = self.agent.calculate_relevance_score(summary, original_text)

        self.assertEqual(relevance_score, expected_relevance_score)

if __name__ == "__main__":
    unittest.main()